requirements.txt
oura-ring
google-cloud-bigquery
